Add compression and encoding outputs.
I'm not too happy with the API.  It mirrors the C++ one, but it's not
particularly Go-ish.  It would be nicer to allow arbitrary compression
and encodings by being able to inject io.Writers/Readers to perform the
transformations or something.  Leaving it as is for the moment so I can
move on.  YAGNI, anyways.