Split keyczar into multiple types w/ embedding
Previously, the main keyczar type handled both key management and was
the base type for all the exposed cryptographic operations.  The new
code has 'keyczar' as the base type for _only_ the key management
functions, and keyCrypter/keySigner as the base types for the
operations.  I've also pulled out the encoding and compression fields
into their own types which are now appropriately embedded.

The keyczar type will have more key management functions added to it as
part of the keyczar-tool (upcoming).